requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==1.3.0
python-dotenv==1.0.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.107.0
pydantic==2.4.2
python-multipart==0.0.6 